backend/user-service/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.0

# Configuration
python-decouple==3.8

# Image Processing
Pillow==10.1.0

# Required for JWT
setuptools
